==> analyses/04_annotate_cell_types.Rmd <==
---
title: Annotate cell types
params:
  input_file: NULL
  output_file: NULL
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.1'
      jupytext_version: 1.2.4
  kernelspec:
    display_name: Python [conda env:.conda-vanderburg_oropharyngeal_cancer]
    language: python
    name: conda-env-.conda-vanderburg_oropharyngeal_cancer-py
---

# Input data and configuration


```{python tags=c("parameters", "hide_input")}
# get default parameters. Either papermill or rmarkdown way.
try:
    input_file = r.params["input_file"]
    output_file = r.params["output_file"]
except:
    print("Could not access params from `r` object. Don't worry if your are running papermill. ")
    input_file = "results/03_correct_data/adata.h5ad"
    output_file = "results/04_annotate_cell-types/adata.h5ad"
```


```{python include=FALSE}
import pandas as pd
import scanpy as sc
import numpy as np
from matplotlib import pyplot as plt
import os
import sys
import gc
import warnings
sys.path.append("lib")
sys.path.append("../lib")
from jupytertools import setwd, fix_logging, display
from toolz.functoolz import pipe, partial
setwd()
fix_logging(sc.settings)
from multiprocessing import Pool
import seaborn as sns
from plotnine import ggplot, aes
import plotnine as n
import scipy.stats as stats
```


```{python}
markers = pd.read_csv("tables/cell_type_markers.csv")
```

```{python}
adata = sc.read_h5ad(input_file)
```

# Determine optimal resolution
We use the Leiden algorithm (Traag et al.) to determine cell-type clusters.

The algorithm depends on a resolution parameter. The higher the resolution, the more clusters will be found. We perform a grid search to test all parameters in a certain range, and hope to find the number of clusters to be stable across a range of resolutions, indicating biologically meaningful clustering.

```{python}
resolutions = np.arange(0.1, 3, 0.05)
```

```{python include=FALSE}
def leiden_with_r(r):
    key = "leiden_{:.3f}".format(r)
    sc.tl.leiden(adata, resolution=r, key_added=key)
    return adata.obs[key]
```

```{python include=FALSE}
p = Pool(32)
```

```{python include=FALSE}
leiden_results = p.map(leiden_with_r, resolutions)
```

```{python include=FALSE}
leiden_results[0].cat.categories.size
```

```{python include=FALSE}
n_clusters = [leiden_results[i].cat.categories.size for i,r in enumerate(resolutions)]
```

```{python include=FALSE}
def smooth(array, length, i):
    return np.mean(array[i:min(i+length-1, len(array)-1)])

clusters_smooth = [smooth(n_clusters, 5, i) for i in range(len(n_clusters))]
```

```{python echo=FALSE}
LEIDEN_RES = 1.1
plt.plot(resolutions, clusters_smooth)
plt.plot(resolutions, n_clusters, color="red")
plt.xlabel("leiden resolution")
plt.ylabel("#clusters")
plt.vlines(x=LEIDEN_RES, ymin=0, ymax=plt.ylim()[1], color="grey")
```

```{python}
sc.tl.leiden(adata, resolution=LEIDEN_RES)
```

There does not seem to be a clear plateau in the curve except the (arguably small) around ~1.0.
The clustering with r=1.0 looks reasonable for assigning cell-types therefore we stick with that for
this task.

```{python}
fig, ax = plt.subplots(figsize=(14, 10))
sc.pl.umap(adata, color="leiden", ax=ax, legend_loc="on data")
```

# Visualize cell-type markers

```{python include=FALSE}
cell_types = np.unique(markers["cell_type"])
```

Perform final clustering with resolution=1:

```{python}
for ct in cell_types:
    marker_genes = markers.loc[markers["cell_type"] == ct,"gene_identifier"]
    sc.pl.umap(adata, color=marker_genes, title=["{}: {}".format(ct, g) for g in marker_genes])
```

## Assign cell types
```{python}
fig, ax = plt.subplots(figsize=(14, 10))
sc.pl.umap(adata, legend_loc="on data", color="leiden", ax=ax)
```

Assign clusters to cell types using the following mapping:

```{python}
annotation = {
    "B cell": [8, 6, 2, 12, 19, 10],
    "CAF": [17],
    "Endothelial cell": [16, 20],
    "Mast cell": [21],
    "NK cell": [5, 23],
    "T cell CD8+": [11, 13, 0, 7],
    "T cell regulatory": [4],
    "T cell CD4+ non-regulatory":[1, 9, 3],
    "myeloid": [14],
    "pDC": [22]
}
```


# Results

```{python include=FALSE}
annot_dict = {
    str(c): ct for ct, clusters in annotation.items() for c in clusters
}
```

```{python include=FALSE}
adata.obs["cell_type"] = [annot_dict.get(c, "unknown") for c in adata.obs["leiden"]]
adata.obs["cell_type_unknown"] = ["known" if ct != "unknown" else ct for ct in adata.obs["cell_type"]]
adata.obs["cell_type_coarse"] = ["T cell" if ct.startswith("T cell") else ct for ct in adata.obs["cell_type"]]
```

```{python}
sc.pl.umap(adata, color=["cell_type_unknown", "cell_type_coarse", "cell_type"])
```

```{python}
display(adata.obs.groupby("cell_type")[["samples"]].count().sort_values("samples"), n=50)
```

```{python include=FALSE}
# fractions by sample
type_per_sample = (adata.obs.groupby(["cell_type", "samples"])
                       .size()
                       .reset_index(name="n_cells")
                       .merge(adata.obs.groupby("samples").size().reset_index(name="n_total_cells"))
                       .assign(frac_cells=lambda x: x["n_cells"]/x["n_total_cells"]))
type_per_sample
```

## Cell-type distribution per sample
```{python echo=FALSE}
(ggplot(type_per_sample, aes(x="samples", y="frac_cells", fill="cell_type"))
     + n.geom_bar(stat='identity')
     + n.scale_fill_brewer(type="qual", palette="Paired")
     + n.theme(subplots_adjust={'right': .4},
               axis_text_x = n.element_text(angle = 90, vjust = 1, hjust=0.5))
)
```

```{python include=FALSE}
# because of https://github.com/pandas-dev/pandas/issues/27519
def t_cell_frac(x):
    return np.sum(x == "T cell")/len(x)

def nk_cell_frac(x):
    return np.sum(x == "NK cell")/len(x)

cell_type_fractions = (adata.obs.groupby(["samples", "facs_purity_cd3", "facs_purity_cd56"])
                       .agg(
                            frac_t_cell=('cell_type_coarse', t_cell_frac),
                            frac_nk_cell=('cell_type_coarse', nk_cell_frac))
                       .dropna()
                       .reset_index())
```

```{python}
display(cell_type_fractions, n=50)
```

## Compare annotations with FACS markers
The correlation between single-cell annotations and FACS is very strong:

```{python echo=FALSE}
x = cell_type_fractions["facs_purity_cd3"]
y = cell_type_fractions["frac_t_cell"]
r, r_p = stats.pearsonr(x, y)
slope, intercept, r_value, p_value, std_err = stats.linregress(x, y)
fig, ax = plt.subplots()
ax.plot(x, y, 'o')
ax.plot(np.array([0,1]), slope*np.array([0,1])+intercept, color="black")
ax.text(x=0, y=1, s="r={:.2f}, p={:.3f}".format(r, r_p))
ax.set_title("T cells: FACS vs. single cell")
ax.set_xlabel("%CD3")
ax.set_ylabel("%T cells")
```

```{python echo=FALSE}
x = cell_type_fractions["facs_purity_cd56"]
y = cell_type_fractions["frac_nk_cell"]
r, r_p = stats.pearsonr(x, y)
slope, intercept, r_value, p_value, std_err = stats.linregress(x, y)
fig, ax = plt.subplots()
ax.plot(x, y, 'o')
ax.plot(np.array([0,.6]), slope*np.array([0,.6])+intercept, color="black")
ax.text(x=0, y=.4, s="r={:.2f}, p={:.3f}".format(r, r_p))
ax.set_title("NK cells: FACS vs. single cell")
ax.set_xlabel("%CD56")
ax.set_ylabel("%NK cells")
```

# Save output

```{python}
adata.write(output_file, compression="lzf")
```

# Summary
The purpose of this notebook is:

* load the normalized and corrected data from the previous step
* Use the Leiden algorithm [Traag et al. 2019](https://www.nature.com/articles/s41598-019-41695-z.pdf) to cluster the single cell data
* Use known marker genes to annotate the clusters


## Results
### UMAP plot colored by celltype

```{python echo=FALSE}
sc.pl.umap(adata, color="cell_type")
```

### cells per cell-type

```{python echo=FALSE}
display(adata.obs.groupby("cell_type")[["samples"]].count().sort_values("samples"), n=50)
```

### cell-distribution per sample

```{python echo=FALSE}
(ggplot(type_per_sample, aes(x="samples", y="frac_cells", fill="cell_type"))
     + n.geom_bar(stat='identity')
     + n.scale_fill_brewer(type="qual", palette="Paired")
     + n.theme(subplots_adjust={'right': .4},
               axis_text_x = n.element_text(angle = 90, vjust = 1, hjust=0.5))
)
```

### Compare annotations with FACS markers:
Overall, the correlation is very strong:

```{python echo=FALSE}
x = cell_type_fractions["facs_purity_cd3"]
y = cell_type_fractions["frac_t_cell"]
r, r_p = stats.pearsonr(x, y)
slope, intercept, r_value, p_value, std_err = stats.linregress(x, y)
fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(14, 5))
ax1.plot(x, y, 'o')
ax1.plot(np.array([0,1]), slope*np.array([0,1])+intercept, color="black")
ax1.text(x=0, y=1, s="r={:.2f}, p={:.3f}".format(r, r_p))
ax1.set_title("T cells: FACS vs. single cell")
ax1.set_xlabel("%CD3")
ax1.set_ylabel("%T cells")

x = cell_type_fractions["facs_purity_cd56"]
y = cell_type_fractions["frac_nk_cell"]
r, r_p = stats.pearsonr(x, y)
slope, intercept, r_value, p_value, std_err = stats.linregress(x, y)
ax2.plot(x, y, 'o')
ax2.plot(np.array([0,.6]), slope*np.array([0,.6])+intercept, color="black")
ax2.text(x=0, y=.4, s="r={:.2f}, p={:.3f}".format(r, r_p))
ax2.set_title("NK cells: FACS vs. single cell")
ax2.set_xlabel("%CD56")
ax2.set_ylabel("%NK cells")

plt.show()
```
