==> analyses/51_analysis_cd39.Rmd <==
---
title: Analysis CD39
params:
  input_file: "../results/50_prepare_analysis_cd39/adata.h5ad"
  input_de_res_dir_cd39: "../results/50-2_run_de_analysis_cd39"
  cpus: 16
---
# Load data

```{r message=FALSE}
library(conflicted)
conflict_prefer("Position", "base")
library(dplyr)
conflict_prefer("filter", "dplyr")
conflict_prefer("select", "dplyr")
library(ggplot2)
library(ggpubr)
library(ggbeeswarm)
library(readr)
library(tidyr)
library(reticulate)
library(clusterProfiler)
library(foreach)
library(doMC)
library(org.Hs.eg.db)
library(cowplot)
library(stringr)
library(kableExtra)
library(EnhancedVolcano)
knitr::knit_engines$set(python = reticulate::eng_python)
reticulate::py_available(TRUE)
# bug in rstudio/reticulate:
matplotlib <- import("matplotlib")
matplotlib$use("Agg", force = TRUE)
registerDoMC(cores=16)
```

```{python}
import scanpy as sc
import matplotlib.pyplot as plt
import matplotlib
import sys
import seaborn as sns

sys.path.extend(("lib", "../lib"))
from jupytertools import *
fix_logging(sc.settings)
matplotlib.rcParams.update({"figure.autolayout": True, "figure.max_open_warning": 0})

adata = sc.read_h5ad(r.params['input_file'])
```

```{r}
obs = py$adata$obs
```

```{r message=FALSE}
get_path_cd39 = function(file_name) {file.path(params$input_de_res_dir_cd39, file_name)}

de_pairwise = read_tsv(get_path_cd39("cd39_status.rda.res.tsv"))
```

```{r export_xlsx}
writexl::write_xlsx(de_pairwise, "pairwise_comparison.xlsx")
```

# DE-analysis
```{python, echo=FALSE}
sc.pl.umap(
    adata,
    color=[
        "cd39_status",
    ],
    size=10,
    ncols=2, 
    legend_loc="on data"
)
```

```{r, include=FALSE}
mapping = data_frame("cluster"=c("cd4_cd8", "cd4_treg", "cd8_treg"), cluster_name=c("CD4+ vs. CD8+", "CD4+ vs. Treg", "CD8+ vs. Treg"))
```

## Differntial expression analysis (pairwise comparison of clusters)
How to read the plot: 
`cd4_cd8` shows genes that are differentially expressed between cells in the CD4 and CD8 cluster. Blue: Enriched in CD8, under-represented in CD4; Orange: Enriched in CD4, uner-represented in CD8. 
Analogous for the other two comparisons. 

```{r echo=FALSE, fig.height=7, fig.width=12}
de_pairwise2 = de_pairwise %>%
    inner_join(mapping) %>%
    filter(str_starts(gene_symbol, "TRAV", negate = TRUE), str_starts(gene_symbol, "TRBV", negate=TRUE)) %>%
    group_by(cluster_name) %>%
    mutate(rk = rank(-abs(logFC))) %>%
    arrange(rk)

p_pairwise = de_pairwise2 %>%
    ungroup() %>%
    filter(rk <= 50) %>%
   # filter(logFC > 0) %>%
    ggplot(aes(x=logFC, y=rk)) +
    geom_vline(xintercept =0, color="grey", linetype="dashed") +
    geom_text(aes(label=gene_symbol, x=-9), angle=0, size=3, nudge_x=1, hjust=0) +
    geom_point(color="black", size=3.5) +
    geom_point(aes(color=logFC), size=3) +
    scale_color_distiller(type="seq", palette="RdYlBu", direction=-1) +
    # geom_text(aes(label="label", x=5, y=10), data=de_genes_count, ha="right", size=8) +
    facet_wrap(~cluster_name, ncol=3) +
    coord_cartesian(xlim=c(-8, 6)) +
    scale_y_reverse() +
    ggtitle(paste0("Pairwise differential gene expression analysis")) +
    theme_bw() +
    theme(legend.position="none")

#ggsave2(filename="figures/de.pdf", plot = p_pairwise, width=12, height=7, units="in")
# ggsave2(filename="figures/de.svg", plot = p_pairwise, width=12, height=7, units="in")


p_pairwise


```


## DE as volcano plot
```{r, fig.width=6, fig.height=6, echo=FALSE, message=FALSE, warning=FALSE}
lapply(mapping$cluster_name, function(comparison) {
  tt = de_pairwise2 %>%
    filter(cluster_name == "CD4+ vs. Treg") 
  
  labelling_cutoff = sort(tt$PValue)[20]
  
  p_cutoff = 0.01 / nrow(tt)
  
  tt %>% 
    EnhancedVolcano(
                x='logFC',
                y='PValue', 
                lab=tt$gene_symbol,
                title=comparison, 
                subtitle="Volcano Plot",
                pLabellingCutoff = labelling_cutoff,
                drawConnectors = TRUE)

})

```
<!-- 
GO-term not really meaningful
## GO-term enrichment analysis
```{r include=FALSE}
MIN_FC = .5
cluster_names = de_pairwise %>% pull(cluster) %>% unique()
universe = de_pairwise %>% pull(gene_symbol) %>% unique()

# dummy see https://github.com/YuLab-SMU/clusterProfiler/issues/207
.x = enrichGO(c("CD8A"), OrgDb = org.Hs.eg.db, keyType = "SYMBOL", ont = "BP", universe = universe)

egos_bp = foreach(clus = cluster_names,
                  .final = function(x) setNames(x, cluster_names)) %dopar% {
  genes = de_pairwise %>%
    filter(cluster == clus, logFC > MIN_FC) %>%
    pull(gene_symbol) %>% unique()
  ego = enrichGO(genes,
         OrgDb = org.Hs.eg.db,
         keyType = "SYMBOL",
         ont = "BP",
         universe = universe
  )
}


egos_mf = foreach(clus = cluster_names,
                  .final = function(x) setNames(x, cluster_names)) %dopar% {
  genes = de_pairwise %>%
    filter(cluster == clus, logFC > MIN_FC) %>%
    pull(gene_symbol) %>% unique()
  ego = enrichGO(genes,
         OrgDb = org.Hs.eg.db,
         keyType = "SYMBOL",
         ont = "MF",
         universe = universe
  )
                  }

egos_bp_inv = foreach(clus = cluster_names,
                  .final = function(x) setNames(x, cluster_names)) %dopar% {
  genes = de_pairwise %>%
    filter(cluster == clus, logFC < -MIN_FC) %>%
    pull(gene_symbol) %>% unique()
  ego = enrichGO(genes,
         OrgDb = org.Hs.eg.db,
         keyType = "SYMBOL",
         ont = "BP",
         universe = universe
  )
}


egos_mf_inv = foreach(clus = cluster_names,
                  .final = function(x) setNames(x, cluster_names)) %dopar% {
  genes = de_pairwise %>%
    filter(cluster == clus, logFC < -MIN_FC) %>%
    pull(gene_symbol) %>% unique()
  ego = enrichGO(genes,
         OrgDb = org.Hs.eg.db,
         keyType = "SYMBOL",
         ont = "MF",
         universe = universe
  )
}

```

```{r include=FALSE, fig.width=8, fig.height=6}
p_go_mf = lapply(names(egos_mf), function(cluster) {
  egos_mf[[cluster]]@result$Description = str_trunc(egos_mf[[cluster]]@result$Description, 50)
  dotplot(egos_mf[[cluster]], showCategory=30, font.size=9) + ggtitle(cluster)
})

p_go_bp = lapply(names(egos_bp), function(cluster) {
  egos_bp[[cluster]]@result$Description = str_trunc(egos_bp[[cluster]]@result$Description, 50)
  dotplot(egos_bp[[cluster]], showCategory=30, font.size=9) + ggtitle(cluster)
})

p_go_mf_inv = lapply(names(egos_mf_inv), function(cluster) {
  egos_mf_inv[[cluster]]@result$Description = str_trunc(egos_mf_inv[[cluster]]@result$Description, 50)
  dotplot(egos_mf_inv[[cluster]], showCategory=30, font.size=9) + ggtitle(cluster)
})

p_go_bp_inv = lapply(names(egos_bp_inv), function(cluster) {
  egos_bp_inv[[cluster]]@result$Description = str_trunc(egos_bp_inv[[cluster]]@result$Description, 50)
  dotplot(egos_bp_inv[[cluster]], showCategory=30, font.size=9) + ggtitle(cluster)
})
```


## Biological process
The plots show the 20 most enriched GO-terms from the biological-process ontology. The further right a point, the
higher the score of the corresponding term.

```{r echo=FALSE, fig.width=18, fig.height=6}
plot_grid(plotlist = p_go_bp, ncol=3) + ggtitle("foo")
plot_grid(plotlist = p_go_bp_inv, ncol=3)
```

## GO term enrichment: Molecular function
The plots show the 20 most enriched GO-terms from the molecular function ontology. The further right a point, the
higher the score of the corresponding term. If fewer terms are shown, there were fewer terms that met the statistical significance threshold.

```{r echo=FALSE, fig.width=18, fig.height=6}
plot_grid(plotlist = p_go_mf, ncol=3)
plot_grid(plotlist = p_go_mf_inv, ncol=3)
```

--> 

## Violin Plots

```{python, include=FALSE}
adata_cd39 = adata[adata.obs["cd39_status"] != "na", :]
genes = ["ENTPD1", "ITGAE", "ITGA1", "PDCD1", "CXCL13", "GZMB", "ALOX5AP", "IFNG", "GNLY", "LAG3", "RUNX2", "FOXP3", "IL2RA", "PIM2", "TNFRSF4", "IL32"]
```

```{python, message=FALSE, warning=FALSE}
sc.pl.dotplot(adata_cd39, groupby="cd39_status", 
                     var_names=genes)

```


```{python, message=FALSE, warning=FALSE}
sc.pl.stacked_violin(adata_cd39, groupby="cd39_status", 
                     var_names=genes, 
                                figsize=(6, 3))
```


```{python, message=FALSE, warning=FALSE, include=FALSE}
sc.pl.stacked_violin(adata_cd39, groupby="cd39_status", 
                     var_names=genes, swap_axes=True, figsize=(1, 12), use_raw=True, save="_narrow.svg")
```

```{python, message=FALSE, warning=FALSE}
sc.pl.stacked_violin(adata_cd39, groupby="cd39_status", 
                     var_names=genes, swap_axes=True, figsize=(5, 12), use_raw=True, save="_wide.svg")
```
